feat: SpeechBallon 수정(DURI-81)
