--- packages/ui/src/assets/Polygon.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-const SvgPolygon = (props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    fill="none"
-    viewBox="0 0 13 10"
-    {...props}
-  >
-    <path fill="#fff" d="M6.5 10 .87.25h11.26z" />
-  </svg>
-);
-export default SvgPolygon;
