--- packages/ui/src/components/SpeechBallon/index.tsx
@@ -1,1 +1,0 @@
-export { SpeechBallon } from "./SpeechBallon";
