+++ packages/ui/src/components/SpeechBallonContainer/index.tsx
@@ -1,0 +1,1 @@
+export { SpeechBallonContainer } from "./SpeechBallonContainer";
