+++ packages/ui/src/components/index.tsx
@@ -6,4 +6,4 @@ export * from './MobileLayout';
 export * from './Navbar';
 export * from './StatusBar';
 export * from './Dropdown';
-export * from './SpeechBallon';
+export * from './SpeechBallonContainer';
